requirements.txt
appnope==0.1.3
asttokens==2.2.1
backcall==0.2.0
cloudpickle==2.2.1
decorator==5.1.1
executing==1.2.0
gym==0.26.2
gym-notices==0.0.8
ipython==8.14.0
jedi==0.18.2
matplotlib-inline==0.1.6
numpy==1.25.1
opencv-python==4.8.0.74
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.39
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.1
pyserial==3.5
six==1.16.0
stack-data==0.6.2
traitlets==5.9.0
wcwidth==0.2.6
